refactor(python): Move _normalize_* from model.py to util/normalize.py
