+++ python/moz/l10n/model.py
@@ -16,9 +16,10 @@ from __future__ import annotations
 
 from collections.abc import Iterator
 from dataclasses import dataclass, field
-from typing import Generic, Iterable, Literal, TypeVar, Union
+from typing import Generic, Literal, TypeVar, Union
 
 from .formats import Format
+from .util.normalize import _normalize_declarations, _normalize_pattern
 
 __all__ = [
     "CatchallKey",
@@ -203,48 +204,6 @@ class SelectMessage:
 
 
 Message = Union[PatternMessage, SelectMessage]
-
-
-def _normalize_pattern(pattern: Pattern, var_refs: set[str]) -> None:
-    i = 0
-    at_str = False
-    while i < len(pattern):
-        el = pattern[i]
-        if isinstance(el, str):
-            if el == "":
-                pattern.pop(i)
-            elif at_str:
-                pattern[i - 1] += el  # type: ignore
-                pattern.pop(i)
-            else:
-                at_str = True
-                i += 1
-        else:
-            at_str = False
-            i += 1
-            for var in el.variable_refs():
-                var_refs.add(var.name)
-
-
-def _normalize_declarations(msg: Message, var_refs: set[str]) -> None:
-    decl_refs = {
-        name: set(var.name for var in decl.variable_refs() if var.name != name)
-        for name, decl in msg.declarations.items()
-    }
-    for name in list(var_refs):
-        var_refs.update(_var_dependencies(decl_refs, name))
-    for name in list(msg.declarations):
-        if name not in var_refs:
-            del msg.declarations[name]
-
-
-def _var_dependencies(decl_refs: dict[str, set[str]], name: str) -> Iterable[str]:
-    drs = decl_refs.get(name, None)
-    if drs:
-        del decl_refs[name]
-        for dr in drs:
-            yield dr
-            yield from _var_dependencies(decl_refs, dr)
 
 
 @dataclass
